requirements.txt
flask
google-generativeai
redis
python-dotenv
faiss-cpu
sentence-transformers
langchain